+++ ext/duckdb/table_function.h
@@ -4,6 +4,7 @@
 struct _rubyDuckDBTableFunction {
     duckdb_table_function table_function;
     VALUE bind_proc;
+    VALUE execute_proc;
 };
 
 typedef struct _rubyDuckDBTableFunction rubyDuckDBTableFunction;